requirements.txt
Flask
python-dotenv
Flask-Bcrypt
waitress
flask_pymongo
